Added Game component
