+++ src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css'
 import styled from 'styled-components'
-import { Board } from './game/conponents/board/Board'
 import FullscreenToggle from './game/conponents/fullScreenToggle';
+import { Game } from './game/conponents/Game/Game';
 
 const PageWrapper = styled.div`
   min-height: 90vh;  
@@ -24,9 +24,8 @@ function App() {
     <PageWrapper>
 
       <FullscreenToggle/>
-      <h1 style={{ color: "black" }}>2048</h1>
-
-      <Board/>
+        <h1 style={{ color: "black" }}>2048</h1>
+      <Game/>
 
       <InfoWrapper>
         {"Game by Inbar and Tal Segal"}
