+++ src/game/conponents/board/Board.tsx
@@ -1,7 +1,5 @@
-import React, { useEffect, useState } from 'react';
 import styled, { css } from 'styled-components';
 import { Tile, TILE_SIZE } from '../tile/Tile';
-import { getNewMatrixByDirection} from '../../logic/boardLogic';
 
 // TILE_SIZE is taken from the the Tile component
 const GRID_SIZE = 4;
@@ -44,165 +42,14 @@ const mapMatrixToTiles = (matrix: number[][]): React.ReactElement[] => {
     return tiles;
 };
 
-export const Board: React.FC = () => {
+interface BoardProps {
+    boardData: number[][];
+}
 
-    const [touchStartX, setTouchStartX] = useState(0);
-    const [touchStartY, setTouchStartY] = useState(0);
-
-    const [boardData, setBoardData] = useState(
-        [
-            [2, 0, 0, 0],
-            [0, 0, 0, 0],
-            [0, 0, 0, 0],
-            [0, 0, 0, 0]
-        ]
-    )
-
-
-    //  let arr = [6, 0, 4, 2, 2, 0,0,5];
-    //  console.log("Before: " + arr + "\nAfter: " + rowFlip(arr));
-
-    // let arr = [0, 4, 2, 2, 0];
-    // console.log("Before: " + arr + "\nAfter: " + rowAfterLeftSwipe(arr));
-
-    // arr = [2, 0, 2, 0, 4];
-    // console.log("Before: " + arr + "\nAfter: " + rowAfterLeftSwipe(arr));
-
-    // arr = [0,4,2,2,8];
-    // console.log("Before: " + arr + "\nAfter: " + rowAfterLeftSwipe(arr));
-
-    // arr = [2,0,4,2,16];
-    // console.log("Before: " + arr + "\nAfter: " + rowAfterLeftSwipe(arr));
-
-    // arr = [0,0,2,0,0];
-    // console.log("Before: " + arr + "\nAfter: " + rowAfterLeftSwipe(arr));
-
-    // arr = [32,2,4,0,2,0];
-    // console.log("Before: " + arr + "\nAfter: " + rowAfterLeftSwipe(arr));
-
-    // arr = [2,0,0,0,0,0,0];
-    // console.log("Before: " + arr + "\nAfter: " + rowAfterLeftSwipe(arr));
-
-    // arr = [2,2,4,8,16,32];
-    // console.log("Before: " + arr + "\nAfter: " + rowAfterLeftSwipe(arr));
-
-    //TESTING for left swipe
-
-
-
-
-    // This is for testing visual of all fonts and colors
-    /*
-        [8,4,2,2],
-        [16,32,64,128],
-        [2048,1024,512,256],
-        [4096,8192,16384,65536]
-    */
-
-    /*
-
-    [0,0,0,2],
-    [0,0,0,0],
-    [0,0,0,0],
-    [0,0,0,0]
-
-    */
-
-    useEffect(() => {
-        const handleKeyDown = (event: KeyboardEvent) => {
-            let newBoardData: number[][];
-
-            switch (event.key) {
-                case 'ArrowLeft':
-                    newBoardData = getNewMatrixByDirection(boardData, "left");
-                    setBoardData(newBoardData);
-                    break;
-                case 'ArrowRight':
-                    newBoardData = getNewMatrixByDirection(boardData, "right");
-                    setBoardData(newBoardData);
-                    break;
-                case 'ArrowUp':
-                    newBoardData = getNewMatrixByDirection(boardData, "up");
-                    setBoardData(newBoardData);
-                    break;
-                case 'ArrowDown':
-                    newBoardData = getNewMatrixByDirection(boardData, "down");
-                    setBoardData(newBoardData);
-                    break;
-                default:
-                    break;
-            }
-        };
-
-        window.addEventListener('keydown', handleKeyDown);
-
-        return () => {
-            window.removeEventListener('keydown', handleKeyDown);
-        }
-    }, [boardData])
-
-
-
-    const handleClick = () => {
-    }
-
-    const handleTouchStart = (e: React.TouchEvent) => {
-        setTouchStartX(e.changedTouches[0].screenX);
-        setTouchStartY(e.changedTouches[0].screenY);
-    }
-
-    const handleTouchEnd = (e: React.TouchEvent) => {
-        const currentX = e.changedTouches[0].screenX;
-        const currentY = e.changedTouches[0].screenY;
-        const deltaX = currentX - touchStartX;
-        const deltaY = currentY - touchStartY;
-
-        const swipeLengthX = Math.abs(deltaX);
-        const swipeLengthY = Math.abs(deltaY);
-
-        let newBoardData: number[][];
-
-        if (swipeLengthX > swipeLengthY) {
-            // There was an X swipe
-            if (swipeLengthX > 50) {
-
-                if (deltaX < 0) // swipe left
-                {
-                    newBoardData = getNewMatrixByDirection(boardData, "left");
-                    setBoardData(newBoardData);
-                }
-                else if (deltaX > 0) {    // swipe right
-                    newBoardData = getNewMatrixByDirection(boardData, "right");
-                    setBoardData(newBoardData);
-                }
-            }
-        }
-        else {
-            // There was a Y swipe
-            if (swipeLengthY > 50) {
-
-                if (deltaY < 0) // swipe up
-                {
-                    newBoardData = getNewMatrixByDirection(boardData, "up");
-                    setBoardData(newBoardData);
-                }
-                else if (deltaY > 0) {    // swipe down
-                    newBoardData = getNewMatrixByDirection(boardData, "down");
-                    setBoardData(newBoardData);
-                }
-            }
-        }
-
-    }
-
+export const Board: React.FC<BoardProps> = ({boardData}) => {
     return (
-        <BoardWrapper
-            onClick={handleClick}
-            onTouchStart={handleTouchStart}
-            onTouchEnd={handleTouchEnd}
-        >
+        <BoardWrapper>
             {mapMatrixToTiles(boardData)}
-
         </BoardWrapper>
     );
 }
